% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs3.R
\name{fit_null_general}
\alias{fit_null_general}
\title{Newton Raphson Function}
\usage{
fit_null_general(init_beta, epsilon, xDats, lt_all, rt_all, k, d)
}
\arguments{
\item{init_beta}{starting values for NR}

\item{epsilon}{stopping criterion for NR}

\item{xDats}{list of design matrices}

\item{lt_all}{n x k matrix of left times}

\item{rt_all}{n x k matrix of right times}

\item{k}{number of outcomes}

\item{d}{number of quadrature points}
}
\description{
Newton Raphson Function
}
