% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs3.R
\name{gen_mICSKAT_dat}
\alias{gen_mICSKAT_dat}
\title{Sample generation}
\usage{
gen_mICSKAT_dat(
  bhFunInv,
  obsTimes = 1:3,
  windowHalf = 0.5,
  n,
  k,
  tauSq,
  gMatCausal,
  effectSizes
)
}
\arguments{
\item{bhFunInv}{A function, the inverse of the baseline hazard function.}

\item{obsTimes}{Vector of the intended observation times.}

\item{windowHalf}{The amount of time before or after the intended obsTimes that a visit might take place.}

\item{n}{number of observations}

\item{k}{number of outcomes}

\item{gMatCausal}{matrix of subsetted genetic information for only a select causal SNPs}

\item{effectSizes}{vector of genetic effects}

\item{tausq}{variance of subject specific random effect}
}
\description{
Sample generation
}
