% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs.R
\name{get_A}
\alias{get_A}
\title{Calculate the denominator term of the derivative of the likelihood}
\usage{
get_A(store, weights, d, n)
}
\arguments{
\item{store}{array of difference between left and right survival values}

\item{weights}{quadrature weights}

\item{d}{number of quadrature nodes}

\item{n}{number of observations}
}
\description{
Calculate the denominator term of the derivative of the likelihood
}
