% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs3.R
\name{multiICSKAT_p_general}
\alias{multiICSKAT_p_general}
\title{test statistic and pvalue calculation}
\usage{
multiICSKAT_p_general(nullFit, xDats, lt_all, rt_all, Itt, a1, a2, G, k, d)
}
\arguments{
\item{nullFit}{beta values fitted from NR}

\item{xDats}{list of design matrices}

\item{lt_all}{n x k matrix of left times}

\item{rt_all}{n x k matrix of right times}

\item{Itt}{information of theta terms from NR}

\item{a1}{first beta distribution shape parameter for weights}

\item{a2}{second beta distribution shape parameter for weights}

\item{G}{n x q matrix of genetic information}

\item{k}{number of outcomes}

\item{d}{number of quadrature points}
}
\description{
test statistic and pvalue calculation
}
