% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs.R
\name{surv_diff}
\alias{surv_diff}
\title{Calculate the difference between left and right survival terms.}
\usage{
surv_diff(l, d, temp_beta, phen, r1, k)
}
\arguments{
\item{l}{observation value of interest}

\item{d}{number of Gaussian Quadrature points}

\item{temp_beta}{vector of parameter estimates}

\item{phen}{list of data matrices containing both left and right information}

\item{r1}{nodes of quadrature points}

\item{k}{number of outcomes}
}
\description{
Calculate the difference between left and right survival terms.
}
