% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs.R
\name{surv_right}
\alias{surv_right}
\title{Calculate the right survival function value.}
\usage{
surv_right(l, d, temp_beta, phen, r1, k)
}
\arguments{
\item{l}{observation value of interest}

\item{d}{number of Gaussian Quadrature points}

\item{temp_beta}{vector of parameter estimates}

\item{phen}{list of data matrices containing both left and right information}

\item{r1}{nodes of quadrature points}

\item{k}{number of outcomes}
}
\description{
Calculate the right survival function value.
}
