% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs.R
\name{without_one_phen}
\alias{without_one_phen}
\title{Calculate the product of the difference between survival terms excluding that of the outcome of interest}
\usage{
without_one_phen(l, k, store)
}
\arguments{
\item{l}{observation value of interest}

\item{k}{number of outcomes}

\item{store}{array of difference between left and right survival values}
}
\description{
Calculate the product of the difference between survival terms excluding that of the outcome of interest
}
