% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIStestfuncs.R
\name{without_two_phen}
\alias{without_two_phen}
\title{Calculate the product of the difference between survival terms excluding that of two outcomes of interest}
\usage{
without_two_phen(l, m, k, store, n, d)
}
\arguments{
\item{l}{observation value of interest}

\item{m}{second observation of interest}

\item{k}{number of outcomes}

\item{store}{array of difference between left and right survival values}

\item{n}{number of observations}

\item{d}{number of quadrature nodes}
}
\value{
A list with the elements:
}
\description{
Calculate the product of the difference between survival terms excluding that of two outcomes of interest
}
